brick/api/endpoints/rings.py
import shutil
import os
import uuid

from pathlib import Path
from pydantic import ValidationError
from celery.result import AsyncResult
from fastapi.responses import JSONResponse
from fastapi import APIRouter,  HTTPException


from ..schemas import BlastRingConfig

router = APIRouter(
    prefix="/rings",
    tags=["ring"],
)


@router.post("/blast")
def create_blast_ring(config: BlastRingConfig):
    # Process the configuration here
    # Example: just return the received config
    return {"received_config": config}