pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "brick"
version = "0.1.0"
authors = [{name="Eike Steinig", email="eike.steinig@unimelb.edu.au"}]
description = "BLAST Ring Image Magick (BRICK) implements BRIG-like figures as API, CLI and REST API stack with Svelte UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer", 
  "biopython", 
  "pandas", 
  "requests", 
  "pre-commit", 
  "fastapi", 
  "celery[redis]", 
  "uvicorn", 
  "pydantic-settings", 
  "python-multipart", 
  "pymongo", 
  "motor"
]

[tool.setuptools]
packages = [
    "brick"
]

[project.scripts]
brick = "brick.terminal:app"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

